+++ src/taskFromEugene3_test/Fractions.java
@@ -1,4 +1,4 @@
-package taskFromEugene3.test;
+package taskFromEugene3_test;
 
 /*3.
 Класс Дробное число со знаком (Fractions). Число должно быть представлено двумя полями:
